First version with bugs
Bug 1: Does not reflect the changes within a collection immediately and needs to be refreshed to reflect the changes. For example, if i delete a link, it does not show within the collection.
Enhancement 1: Need to allow users to edit the Collection names
Enhancement 2: Need to allow users to create a collection without needing to save all Tabs.